fix multiple-sub handling
assume you have the following sub setup:
subs:
  rhel:
    - abcd
  epel:
    - def0
  zork:
    - ffff
and machine has rhel and epel already attached.

When running the script, you get:
 subscription abcd for rhel product is already attached
 subscription def0 for rhel product is already attached
 subscription abcd for epel product is already attached
 subscription def0 for epel product is already attached
 subscription abcd for zork product is already attached
 subscription def0 for zork product is already attached

now lines 2 and 3 are clearly wrong. there is no def0 for rhel and no
abcd for epel. so are lines 5 and 6 as zork is not attached at all.

the issue was that the cp_id was checked against ALL desired subs
not the current one